Switch I2C slave to 16-bit sub-addressing (EEPROM-style)
Replace 8-bit register address with 16-bit sub-addressing compatible
with 24C256/24C512 EEPROM protocol, giving 64K register address space
for future automotive remote-display features.

Protocol: [slave+W] [addr_hi] [addr_lo] [data...]
Address auto-increments and wraps at 0xFFFF -> 0x0000.
Current-address read (no prior address set) uses last set address.

Register map reorganized:
  0x0000-0x00FF  Device info (FW version BCD at 0x0000-0x0001)
  0x0100-0x01FF  Status (DIP switches at 0x0100)
  0x0200-0x02FF  Control (LED at 0x0200)

HAL state machine: IDLE -> ADDR_HI -> ADDR_LO -> RECEIVING_DATA
Callbacks changed from uint8 to uint16 register address.

Tested on 983HH: FW version read, LED control, DIP read all
verified via i2ctransfer on Pi4.

+++ hal/hal_riic_slave.h
@@ -1,5 +1,12 @@
 /*
  * hal_riic_slave.h - RIIC0 hardware I2C slave driver (interrupt-driven)
+ *
+ * 16-bit sub-addressing (EEPROM-style, 24C256/24C512 compatible):
+ *   Write: [slave+W] [addr_hi] [addr_lo] [data0] [data1] ...
+ *   Read:  [slave+W] [addr_hi] [addr_lo] [slave+R] [data0] ...
+ *   Current-address read: [slave+R] [data0] ... (continues from last addr)
+ *
+ * Address auto-increments after each byte and wraps at 0xFFFF -> 0x0000.
  */
 
 #ifndef HAL_RIIC_SLAVE_H
@@ -7,9 +14,9 @@
 
 #include "dr7f701686.dvf.h"
 
-/* Application callbacks for I2C slave register access */
-typedef void    (*hal_riic_slave_write_cb)(uint8 reg_addr, uint8 value);
-typedef uint8   (*hal_riic_slave_read_cb)(uint8 reg_addr);
+/* Application callbacks for I2C slave register access (16-bit address) */
+typedef void    (*hal_riic_slave_write_cb)(uint16 reg_addr, uint8 value);
+typedef uint8   (*hal_riic_slave_read_cb)(uint16 reg_addr);
 
 /*
  * Initialize RIIC0 as I2C slave.
